refactor: share stats accumulation
